+++ python/apollo_openapi/apis/paths/openapi_v1_apollo_audit_properties.py
@@ -1,7 +1,0 @@
-from apollo_openapi.paths.openapi_v1_apollo_audit_properties.get import ApiForget
-
-
-class OpenapiV1ApolloAuditProperties(
-    ApiForget,
-):
-    pass