+++ python/apollo_openapi/apis/paths/openapi_v1_apps_app_id_appnamespaces_namespace_name_usage.py
@@ -1,7 +1,0 @@
-from apollo_openapi.paths.openapi_v1_apps_app_id_appnamespaces_namespace_name_usage.get import ApiForget
-
-
-class OpenapiV1AppsAppIdAppnamespacesNamespaceNameUsage(
-    ApiForget,
-):
-    pass