+++ python/apollo_openapi/apis/paths/openapi_v1_apps_app_id_envs_env_clusters_cluster_name_export.py
@@ -1,9 +1,0 @@
-from apollo_openapi.paths.openapi_v1_apps_app_id_envs_env_clusters_cluster_name_export.get import ApiForget
-from apollo_openapi.paths.openapi_v1_apps_app_id_envs_env_clusters_cluster_name_export.head import ApiForhead
-
-
-class OpenapiV1AppsAppIdEnvsEnvClustersClusterNameExport(
-    ApiForget,
-    ApiForhead,
-):
-    pass