+++ python/apollo_openapi/apis/paths/openapi_v1_apps_app_id_envs_env_clusters_cluster_name_import.py
@@ -1,7 +1,0 @@
-from apollo_openapi.paths.openapi_v1_apps_app_id_envs_env_clusters_cluster_name_import.post import ApiForpost
-
-
-class OpenapiV1AppsAppIdEnvsEnvClustersClusterNameImport(
-    ApiForpost,
-):
-    pass