+++ python/apollo_openapi/apis/paths/openapi_v1_apps_app_id_envs_env_namespaces_namespace_name_role_users.py
@@ -1,7 +1,0 @@
-from apollo_openapi.paths.openapi_v1_apps_app_id_envs_env_namespaces_namespace_name_role_users.get import ApiForget
-
-
-class OpenapiV1AppsAppIdEnvsEnvNamespacesNamespaceNameRoleUsers(
-    ApiForget,
-):
-    pass