+++ python/apollo_openapi/apis/paths/openapi_v1_apps_app_id_namespaces_namespace_name_permissions_permission_type.py
@@ -1,7 +1,0 @@
-from apollo_openapi.paths.openapi_v1_apps_app_id_namespaces_namespace_name_permissions_permission_type.get import ApiForget
-
-
-class OpenapiV1AppsAppIdNamespacesNamespaceNamePermissionsPermissionType(
-    ApiForget,
-):
-    pass