+++ python/apollo_openapi/apis/paths/openapi_v1_apps_app_id_roles_role_type.py
@@ -1,9 +1,0 @@
-from apollo_openapi.paths.openapi_v1_apps_app_id_roles_role_type.post import ApiForpost
-from apollo_openapi.paths.openapi_v1_apps_app_id_roles_role_type.delete import ApiFordelete
-
-
-class OpenapiV1AppsAppIdRolesRoleType(
-    ApiForpost,
-    ApiFordelete,
-):
-    pass