+++ python/apollo_openapi/apis/paths/openapi_v1_envs_env_apps_app_id_clusters_cluster_name_namespaces_namespace_name.py
@@ -1,7 +1,0 @@
-from apollo_openapi.paths.openapi_v1_envs_env_apps_app_id_clusters_cluster_name_namespaces_namespace_name.get import ApiForget
-
-
-class OpenapiV1EnvsEnvAppsAppIdClustersClusterNameNamespacesNamespaceName(
-    ApiForget,
-):
-    pass