+++ python/apollo_openapi/apis/paths/openapi_v1_envs_env_instances_by_namespace_and_releases_not_in.py
@@ -1,7 +1,0 @@
-from apollo_openapi.paths.openapi_v1_envs_env_instances_by_namespace_and_releases_not_in.get import ApiForget
-
-
-class OpenapiV1EnvsEnvInstancesByNamespaceAndReleasesNotIn(
-    ApiForget,
-):
-    pass