+++ python/apollo_openapi/apis/paths/openapi_v1_envs_env_instances_by_release.py
@@ -1,7 +1,0 @@
-from apollo_openapi.paths.openapi_v1_envs_env_instances_by_release.get import ApiForget
-
-
-class OpenapiV1EnvsEnvInstancesByRelease(
-    ApiForget,
-):
-    pass