+++ python/apollo_openapi/apis/paths/openapi_v1_envs_env_releases_comparison.py
@@ -1,7 +1,0 @@
-from apollo_openapi.paths.openapi_v1_envs_env_releases_comparison.get import ApiForget
-
-
-class OpenapiV1EnvsEnvReleasesComparison(
-    ApiForget,
-):
-    pass