+++ python/apollo_openapi/apis/paths/openapi_v1_envs_env_releases_release_id.py
@@ -1,7 +1,0 @@
-from apollo_openapi.paths.openapi_v1_envs_env_releases_release_id.get import ApiForget
-
-
-class OpenapiV1EnvsEnvReleasesReleaseId(
-    ApiForget,
-):
-    pass