+++ python/apollo_openapi/apis/paths/openapi_v1_envs_env_releases_release_id_rollback.py
@@ -1,7 +1,0 @@
-from apollo_openapi.paths.openapi_v1_envs_env_releases_release_id_rollback.put import ApiForput
-
-
-class OpenapiV1EnvsEnvReleasesReleaseIdRollback(
-    ApiForput,
-):
-    pass