+++ python/apollo_openapi/apis/paths/openapi_v1_page_settings.py
@@ -1,7 +1,0 @@
-from apollo_openapi.paths.openapi_v1_page_settings.get import ApiForget
-
-
-class OpenapiV1PageSettings(
-    ApiForget,
-):
-    pass