+++ python/apollo_openapi/apis/paths/openapi_v1_system_info.py
@@ -1,7 +1,0 @@
-from apollo_openapi.paths.openapi_v1_system_info.get import ApiForget
-
-
-class OpenapiV1SystemInfo(
-    ApiForget,
-):
-    pass