+++ python/apollo_openapi/apis/paths/openapi_v1_system_role_manage_app_master.py
@@ -1,7 +1,0 @@
-from apollo_openapi.paths.openapi_v1_system_role_manage_app_master.get import ApiForget
-
-
-class OpenapiV1SystemRoleManageAppMaster(
-    ApiForget,
-):
-    pass