+++ python/apollo_openapi/apis/paths/openapi_v1_users_enabled.py
@@ -1,7 +1,0 @@
-from apollo_openapi.paths.openapi_v1_users_enabled.put import ApiForput
-
-
-class OpenapiV1UsersEnabled(
-    ApiForput,
-):
-    pass